Update `powsybl-dependencies` to version `2025.0.2` and refactor CGMES control area handling (#72)
* Refactor CGMES control area handling and update dependencies.

Replaced `CgmesControlArea` with `Area` and simplified related logic to improve maintainability. Updated `powsybl-dependencies` to version `2025.0.0` in `pom.xml` for compatibility with the latest library changes.

* Refactor area mapping to use method reference syntax.

Replaced lambda expression with a method reference to improve readability and adhere to clean code principles. This change does not alter functionality but simplifies the stream operation.

* Update powsybl-dependencies to version 2025.0.2 in pom.xml.

Signed-off-by: Hugo SCHINDLER <hugo.schindler@rte-france.com>